Pointer to `long` conversion warnings due to mismatched size on some platforms

+++ graphics_mmx.cpp
@@ -26,6 +26,8 @@
 
 #ifdef USE_X86_GFX
 
+#include <stdint.h>
+
 #include <mmintrin.h>
 #include <math.h>
 #ifndef M_PI
@@ -41,7 +43,7 @@ void imageFilterMean_MMX(unsigned char *src1, unsigned char *src2, unsigned char
     int n = length;
 
     // Compute first few values so we're on a 8-byte boundary in dst
-    while( (((long)dst & 0x7) > 0) && (n > 0) ) {
+    while( (((uintptr_t)dst & 0x7) > 0) && (n > 0) ) {
         MEAN_PIXEL();
         --n; ++dst; ++src1; ++src2;
     }
@@ -73,7 +75,7 @@ void imageFilterAddTo_MMX(unsigned char *dst, unsigned char *src, int length)
     int n = length;
 
     // Compute first few values so we're on a 8-byte boundary in dst
-    while( (((long)dst & 0x7) > 0) && (n > 0) ) {
+    while( (((uintptr_t)dst & 0x7) > 0) && (n > 0) ) {
         ADDTO_PIXEL();
         --n; ++dst; ++src;
     }
@@ -99,7 +101,7 @@ void imageFilterSubFrom_MMX(unsigned char *dst, unsigned char *src, int length)
     int n = length;
 
     // Compute first few values so we're on a 8-byte boundary in dst
-    while( (((long)dst & 0x7) > 0) && (n > 0) ) {
+    while( (((uintptr_t)dst & 0x7) > 0) && (n > 0) ) {
         SUBFROM_PIXEL();
         --n; ++dst; ++src;
     }
